Modify the implementation of Supporting multiple Hbase secure clusters
