+++ core/src/main/scala/org/apache/spark/sql/execution/datasources/hbase/HBaseRelation.scala
@@ -31,9 +31,8 @@ import org.apache.hadoop.hbase.io.ImmutableBytesWritable
 import org.apache.hadoop.hbase.mapreduce.TableOutputFormat
 import org.apache.hadoop.hbase.util.Bytes
 import org.apache.hadoop.hbase._
-import org.apache.hadoop.io.DataOutputBuffer
 import org.apache.hadoop.mapreduce.Job
-import org.apache.hadoop.security.{Credentials, UserGroupInformation}
+import org.apache.hadoop.security.UserGroupInformation
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.sources._
 import org.apache.spark.sql.types._
@@ -117,15 +116,12 @@ case class HBaseRelation(
 
   def hbaseConf = wrappedConf.value
 
-  val serializedCredentials = {
-    if (HBaseCredentialsManager.manager.isCredentialsRequired(hbaseConf)) {
-      val credentials = HBaseCredentialsManager.manager.getCredentialsForCluster(hbaseConf)
-      UserGroupInformation.getCurrentUser.addCredentials(credentials)
-      HBaseRelation.serialize(credentials)
+  val serializedCredentials =
+    if (SHCCredentialsManager.manager.isCredentialsRequired(hbaseConf)) {
+      SHCCredentialsManager.manager.getCredentialsForCluster(hbaseConf)
     } else {
       null
     }
-  }
 
   def createTable() {
     if (catalog.numReg > 3) {
@@ -236,8 +232,9 @@ case class HBaseRelation(
 
     rdd.mapPartitions(iter => {
       if (null != serializedCredentials) {
-        UserGroupInformation.getCurrentUser
-          .addCredentials(HBaseRelation.deserialize(serializedCredentials))
+        val creds = SHCCredentialsManager.deserialize(serializedCredentials)
+        SHCCredentialsManager.addLogs("Task", creds) // for debug
+        UserGroupInformation.getCurrentUser.addCredentials(creds)
       }
       iter.map(convertToPut)
     }).saveAsNewAPIHadoopDataset(job.getConfiguration)
@@ -328,28 +325,4 @@ object HBaseRelation {
   val HBASE_CONFIGURATION = "hbaseConfiguration"
   // HBase configuration file such as HBase-site.xml, core-site.xml
   val HBASE_CONFIGFILE = "hbaseConfigFile"
-
-  def serialize(credentials: Credentials): Array[Byte] = {
-    if (credentials != null) {
-      val dob = new DataOutputBuffer()
-      credentials.writeTokenStorageToStream(dob)
-      val dobCopy = new Array[Byte](dob.getLength)
-      System.arraycopy(dob.getData, 0, dobCopy, 0, dobCopy.length)
-      dobCopy
-    } else {
-      null
-    }
-  }
-
-  def deserialize(credsBytes: Array[Byte]): Credentials = {
-    if (credsBytes != null) {
-      val byteStream = new ByteArrayInputStream(credsBytes)
-      val dataStream = new DataInputStream(byteStream)
-      val credentials = new Credentials()
-      credentials.readTokenStorageStream(dataStream)
-      credentials
-    } else {
-      null
-    }
-  }
 }
