+++ core/src/main/scala/org/apache/spark/sql/execution/datasources/hbase/HBaseTableScan.scala
@@ -258,7 +258,9 @@ private[hbase] class HBaseTableScanRDD(
 
   override def compute(split: Partition, context: TaskContext): Iterator[Row] = {
     if (null != credentials) {
-      UserGroupInformation.getCurrentUser.addCredentials(HBaseRelation.deserialize(credentials))
+      val creds = SHCCredentialsManager.deserialize(credentials)
+      SHCCredentialsManager.addLogs("Task", creds) // for debug
+      UserGroupInformation.getCurrentUser.addCredentials(creds)
     }
     val ord = hbase.ord//implicitly[Ordering[HBaseType]]
     val partition = split.asInstanceOf[HBaseScanPartition]
